Refactor Automation and Foreman Controllers
