+++ app/controllers/application_controller/filter.rb
@@ -420,9 +420,9 @@ module ApplicationController::Filter
     end
 
     if ["delete", "saveit"].include?(params[:button])
-      if @edit[:in_explorer] || x_active_tree == :storage_tree
-        if "cs_filter_tree" == x_active_tree.to_s
-          build_configuration_manager_tree(:cs_filter, x_active_tree)
+      if @edit[:in_explorer]
+        if "configuration_manager_cs_filter_tree" == x_active_tree.to_s
+          build_configuration_manager_tree(:configuration_manager_cs_filter, x_active_tree)
         else
           tree_type = x_active_tree.to_s.sub(/_tree/, '').to_sym
           builder = TreeBuilder.class_for_type(tree_type)
@@ -451,7 +451,7 @@ module ApplicationController::Filter
       if ["delete", "saveit"].include?(params[:button])
         if @edit[:in_explorer] || x_active_tree == :storage_tree
           tree_name = x_active_tree.to_s
-          if "cs_filter_tree" == tree_name
+          if "configuration_manager_cs_filter_tree" == tree_name
             page.replace_html("#{tree_name}_div", :partial => "provider_foreman/#{tree_name}")
           else
             page.replace_html("#{tree_name}_div", :partial => "shared/tree", :locals => {
@@ -511,7 +511,7 @@ module ApplicationController::Filter
         if x_active_tree.to_s =~ /_filter_tree$/ &&
            !["Vm", "MiqTemplate"].include?(TreeBuilder.get_model_for_prefix(@nodetype))
           search_id = 0
-          if x_active_tree == :cs_filter_tree
+          if x_active_tree == :configuration_manager_cs_filter_tree
             adv_search_build("ConfiguredSystem")
           else
             adv_search_build(vm_model_from_active_tree(x_active_tree))
