+++ app/controllers/mixins/sandbox.rb
@@ -50,7 +50,8 @@ module Sandbox
     condition_tree
     containers_tree
     containers_filter_tree
-    cs_filter_tree
+    configuration_manager_cs_filter_tree
+    configuration_manager_providers_tree
     customization_templates_tree
     datacenter_tree
     datastore_tree
@@ -122,7 +123,8 @@ module Sandbox
     condition
     containers
     containers_filter
-    cs_filter
+    configuration_manager_cs_filter
+    configuration_manager_providers
     customization_templates
     datastores
     db
