+++ app/controllers/provider_foreman_controller.rb
@@ -6,6 +6,7 @@ class ProviderForemanController < ApplicationController
   after_action :set_session_data
 
   include Mixins::GenericSessionMixin
+  include Mixins::ManagerControllerMixin
 
   def self.model
     ManageIQ::Providers::ConfigurationManager
@@ -27,6 +28,10 @@ class ProviderForemanController < ApplicationController
     end
   end
 
+  def manager_prefix
+    'configuration_manager'
+  end
+
   def model_to_name(provmodel)
     ProviderForemanController.model_to_name(provmodel)
   end
@@ -35,18 +40,10 @@ class ProviderForemanController < ApplicationController
     ProviderForemanController.model_to_type_name(provmodel)
   end
 
-  def index
-    redirect_to :action => 'explorer'
-  end
-
-  def show_list
-    redirect_to :action => 'explorer', :flash_msg => @flash_array.try(:fetch_path, 0, :message)
-  end
-
   def new
     assert_privileges("provider_foreman_add_provider")
-    @provider_cfgmgmt = ManageIQ::Providers::ConfigurationManager.new
-    @provider_types = ["Ansible Tower", ui_lookup(:ui_title => 'foreman')]
+    @provider_manager = ManageIQ::Providers::ConfigurationManager.new
+    @provider_types = [ui_lookup(:ui_title => 'foreman')]
     @server_zones = Zone.in_my_region.order('lower(description)').pluck(:description, :name)
     render_form
   end
@@ -56,15 +53,15 @@ class ProviderForemanController < ApplicationController
     @server_zones = Zone.in_my_region.order('lower(description)').pluck(:description, :name)
     case params[:button]
     when "cancel"
-      cancel_provider_foreman
+      cancel_provider
     when "save"
-      add_provider_foreman
-      save_provider_foreman
+      add_provider
+      save_provider
     else
       assert_privileges("provider_foreman_edit_provider")
       manager_id            = from_cid(params[:miq_grid_checks] || params[:id] || find_checked_items[0])
-      @provider_cfgmgmt     = find_record(ManageIQ::Providers::ConfigurationManager, manager_id)
-      @providerdisplay_type = model_to_name(@provider_cfgmgmt.type)
+      @provider_manager     = find_record(ManageIQ::Providers::ConfigurationManager, manager_id)
+      @providerdisplay_type = model_to_name(@provider_manager.type)
       render_form
     end
   end
@@ -79,7 +76,7 @@ class ProviderForemanController < ApplicationController
     else
       providers.each do |provider|
         AuditEvent.success(
-          :event        => "configuration_manager_record_delete_initiated", # TODO: Should be provider_record_delete_initiated
+          :event        => "provider_record_delete_initiated",
           :message      => "[#{provider.name}] Record delete initiated",
           :target_id    => provider.id,
           :target_class => provider.type,
@@ -104,7 +101,7 @@ class ProviderForemanController < ApplicationController
 
   def provision
     assert_privileges("provider_foreman_configured_system_provision") if x_active_accord == :configuration_manager_providers
-    assert_privileges("configured_system_provision") if x_active_accord == :cs_filter
+    assert_privileges("configured_system_provision") if x_active_accord == :configuration_manager_cs_filter
     provisioning_ids = find_checked_items
     provisioning_ids.push(params[:id]) if provisioning_ids.empty?
 
@@ -133,68 +130,11 @@ class ProviderForemanController < ApplicationController
     when :configuration_manager_providers
       assert_privileges("provider_foreman_configured_system_tag")
       tagging_edit('ConfiguredSystem', false)
-    when :cs_filter
+    when :configuration_manager_cs_filter
       assert_privileges("configured_system_tag")
       tagging_edit('ConfiguredSystem', false)
     end
     render_tagging_form
-  end
-
-  def add_provider_foreman
-    find_or_build_provider
-    sync_form_to_instance
-
-    update_authentication_provider(:save)
-  end
-
-  def update_authentication_provider(mode = :validate)
-    @provider_cfgmgmt.update_authentication(build_credentials, :save => mode == :save)
-  end
-
-  def build_credentials
-    return {} unless params[:log_userid]
-    {
-      :default => {
-        :userid   => params[:log_userid],
-        :password => params[:log_password] || @provider_cfgmgmt.authentication_password
-      }
-    }
-  end
-
-  def save_provider_foreman
-    if @provider_cfgmgmt.save
-      construct_edit
-      AuditEvent.success(build_created_audit(@provider_cfgmgmt, @edit))
-      @in_a_form = false
-      @sb[:action] = nil
-      model = "#{model_to_name(@provider_cfgmgmt.type)} #{ui_lookup(:model => 'ExtManagementSystem')}"
-      if params[:id] == "new"
-        add_flash(_("%{model} \"%{name}\" was added") % {:model => model, :name => @provider_cfgmgmt.name})
-        process_managers([@provider_cfgmgmt.configuration_manager.id], "refresh_ems")
-      else
-        add_flash(_("%{model} \"%{name}\" was updated") % {:model => model, :name => @provider_cfgmgmt.name})
-      end
-      replace_right_cell(:replace_trees => [:configuration_manager_providers])
-    else
-      @provider_cfgmgmt.errors.each do |field, msg|
-        @sb[:action] = nil
-        add_flash("#{field.to_s.capitalize} #{msg}", :error)
-      end
-      render_flash
-    end
-  end
-
-  def cancel_provider_foreman
-    @in_a_form = false
-    @sb[:action] = nil
-    if params[:id] == "new"
-      add_flash(_("Add of Configuration Manager %{provider} was cancelled by the user") %
-        {:provider => ui_lookup(:model => 'ExtManagementSystem')})
-    else
-      add_flash(_("Edit of Configuration Manager %{provider} was cancelled by the user") %
-        {:provider => ui_lookup(:model => 'ExtManagementSystem')})
-    end
-    replace_right_cell
   end
 
   def provider_foreman_form_fields
@@ -204,10 +144,10 @@ class ProviderForemanController < ApplicationController
                              :zone => Zone.in_my_region.size >= 1 ? Zone.in_my_region.first.name : nil
                            } if params[:id] == "new"
 
-    config_mgr = find_record(ManageIQ::Providers::ConfigurationManager, params[:id])
-    provider   = config_mgr.provider
+    manager = find_record(ManageIQ::Providers::ConfigurationManager, params[:id])
+    provider   = manager.provider
 
-    render :json => {:provtype   => model_to_name(config_mgr.type),
+    render :json => {:provtype   => model_to_name(manager.type),
                      :name       => provider.name,
                      :zone       => provider.zone.name,
                      :url        => provider.url,
@@ -215,72 +155,8 @@ class ProviderForemanController < ApplicationController
                      :log_userid => provider.authentications.first.userid}
   end
 
-  def authentication_validate
-    find_or_build_provider
-    sync_form_to_instance
-    update_authentication_provider
-
-    begin
-      @provider_cfgmgmt.verify_credentials(params[:type])
-    rescue => error
-      render_flash(_("Credential validation was not successful: %{details}") % {:details => error}, :error)
-    else
-      render_flash(_("Credential validation was successful"))
-    end
-  end
-
-  def show(id = nil)
-    @flash_array = [] if params[:display]
-    @sb[:action] = nil
-
-    @display = params[:display] || "main"
-    @lastaction = "show"
-    @showtype = "config"
-    @record = if configuration_profile_record?
-                find_record(ConfigurationProfile, id || params[:id])
-              else
-                find_record(ConfiguredSystem, id || params[:id])
-              end
-    return if record_no_longer_exists?(@record)
-
-    @explorer = true if request.xml_http_request? # Ajax request means in explorer
-
-    @gtl_url = "/show"
-  end
-
-  def tree_select
-    @lastaction = "explorer"
-    @flash_array = nil
-    self.x_active_tree = params[:tree] if params[:tree]
-    self.x_node = params[:id]
-    load_or_clear_adv_search
-    apply_node_search_text if x_active_tree == :configuration_manager_providers_tree
-
-    unless action_name == "reload"
-      @sb[:active_tab] = if active_tab_configured_systems?
-                           'configured_systems'
-                         else
-                           'summary'
-                         end
-      replace_right_cell
-    else
-      replace_right_cell(:replace_trees => [:configuration_manager_providers])
-    end
-  end
-
-  def accordion_select
-    @lastaction = "explorer"
-
-    @sb[:foreman_search_text] ||= {}
-    @sb[:foreman_search_text]["#{x_active_accord}_search_text"] = @search_text
-
-    self.x_active_accord = params[:id].sub(/_accord$/, '')
-    self.x_active_tree   = "#{x_active_accord}_tree"
-
-    @search_text = @sb[:foreman_search_text]["#{x_active_accord}_search_text"]
-
-    load_or_clear_adv_search
-    replace_right_cell(:replace_trees => [x_active_accord])
+  def managed_group_kls
+    ConfigurationProfile
   end
 
   def load_or_clear_adv_search
@@ -288,12 +164,12 @@ class ProviderForemanController < ApplicationController
     session[:edit] = @edit
     @explorer = true
 
-    if x_active_tree != :cs_filter_tree || x_node == "root"
+    if x_active_tree != :configuration_manager_cs_filter_tree || x_node == "root"
       listnav_search_selected(0)
     else
       @nodetype, id = parse_nodetype_and_id(valid_active_node(x_node))
 
-      if x_active_tree == :cs_filter_tree && @nodetype == "xx-csf"
+      if x_active_tree == :configuration_manager_cs_filter_tree
         search_id = @nodetype == "root" ? 0 : from_cid(id)
         listnav_search_selected(search_id) unless params.key?(:search_text) # Clear or set the adv search filter
         if @edit[:adv_search_applied] &&
@@ -321,7 +197,7 @@ class ProviderForemanController < ApplicationController
     @record =
       case x_active_tree
       when :configuration_manager_providers_tree then configuration_manager_providers_tree_rec
-      when :cs_filter_tree                       then cs_filter_tree_rec
+      when :configuration_manager_cs_filter_tree then cs_filter_tree_rec
       end
   end
 
@@ -358,7 +234,7 @@ class ProviderForemanController < ApplicationController
     end
   end
 
-  def cs_filter_tree_rec
+  def configuration_manager_cs_filter_tree_rec
     nodes = x_node.split('-')
     case nodes.first
     when "root", "xx" then find_record(ConfiguredSystem, params[:id])
@@ -369,7 +245,6 @@ class ProviderForemanController < ApplicationController
   def show_record(_id = nil)
     @display    = params[:display] || "main" unless pagination_or_gtl_request?
     @lastaction = "show"
-    @showtype   = "config"
 
     if @record.nil?
       add_flash(_("Error: Record no longer exists in the database"), :error)
@@ -385,95 +260,22 @@ class ProviderForemanController < ApplicationController
     return unless @display == 'main'
     @showtype     = "main"
     @button_group = case x_active_accord
-                    when :cs_filter
+                    when :configuration_manager_cs_filter
                       rec_cls.to_s
                     when :configuration_manager_providers
                       "provider_foreman_#{rec_cls}"
                     end
   end
 
-  def explorer
-    @explorer = true
-    @lastaction = "explorer"
-
-    # if AJAX request, replace right cell, and return
-    if request.xml_http_request?
-      replace_right_cell
-      return
-    end
-
-    if params[:accordion]
-      self.x_active_tree   = "#{params[:accordion]}_tree"
-      self.x_active_accord = params[:accordion]
-    end
-    if params[:button]
-      @miq_after_onload = "miqAjax('/#{controller_name}/x_button?pressed=#{params[:button]}');"
-    end
-
-    build_accordions_and_trees
-
-    params.instance_variable_get(:@parameters).merge!(session[:exp_parms]) if session[:exp_parms]  # Grab any explorer parm overrides
-    session.delete(:exp_parms)
-    @in_a_form = false
-
-    if params[:id] # If a tree node id came in, show in one of the trees
-      nodetype, id = params[:id].split("-")
-      # treebuilder initializes x_node to root first time in locals_for_render,
-      # need to set this here to force & activate node when link is clicked outside of explorer.
-      @reselect_node = self.x_node = "#{nodetype}-#{to_cid(id)}"
-      get_node_info(x_node)
-    end
-    render :layout => "application"
-  end
-
-  def tree_autoload
-    @view ||= session[:view]
-    super
-  end
-
-  def change_tab
-    @sb[:active_tab] = params[:tab_id]
-    replace_right_cell
-  end
-
-  def cs_form_field_changed
-    id = params[:id]
-    return unless load_edit("cs_edit__#{id}", "replace_cell__explorer")
-    cs_edit_get_form_vars
-    render :update do |page|
-      page << javascript_prologue
-      page << javascript_hide("buttons_off")
-      page << javascript_show("buttons_on")
-    end
-  end
-
-  private ###########
-
-  def display_node(id, model)
-    if @record.nil?
-      self.x_node = "root"
-      get_node_info("root")
-    else
-      show_record(from_cid(id))
-      model_string = ui_lookup(:model => @record.class.to_s)
-      @right_cell_text = _("%{model} \"%{name}\"") % {:name => @record.name, :model => model_string}
-    end
-  end
+ private ###########
 
   def find_or_build_provider
-    @provider_cfgmgmt   = provider_class_from_provtype.new if params[:id] == "new"
-    @provider_cfgmgmt ||= find_record(ManageIQ::Providers::ConfigurationManager, params[:id]).provider # TODO: Why is params[:id] an ExtManagementSystem ID instead of Provider ID?
+    @provider = provider_class_from_provtype.new if params[:id] == "new"
+    @provider ||= find_record(ManageIQ::Providers::ConfigurationManager, params[:id]).provider # TODO: Why is params[:id] an ExtManagementSystem ID instead of Provider ID?
   end
 
   def provider_class_from_provtype
     params[:provtype] = ManageIQ::Providers::Foreman::Provider
-  end
-
-  def sync_form_to_instance
-    @provider_cfgmgmt.name       = params[:name]
-    @provider_cfgmgmt.url        = params[:url]
-    @provider_cfgmgmt.verify_ssl = params[:verify_ssl].eql?("on")
-    @provider_cfgmgmt.zone       = Zone.find_by_name(params[:zone].to_s)
   end
 
   def features
@@ -483,7 +285,7 @@ class ProviderForemanController < ApplicationController
       :title    => _("Providers")},
      {:role     => "configured_systems_filter_accord",
       :role_any => true,
-      :name     => :cs_filter,
+      :name     => :configuration_manager_cs_filter,
       :title    => _("Configured Systems")},
      ].map do |hsh|
       ApplicationController::Feature.new_with_hash(hsh)
@@ -494,7 +296,7 @@ class ProviderForemanController < ApplicationController
     tree = case name
            when :configuration_manager_providers_tree
              TreeBuilderConfigurationManager.new(name, type, @sb)
-           when :cs_filter_tree
+           when :configuration_manager_cs_filter_tree
              TreeBuilderConfigurationManagerConfiguredSystems.new(name, type, @sb)
            end
     instance_variable_set :"@#{name}", tree.tree_nodes
@@ -527,7 +329,7 @@ class ProviderForemanController < ApplicationController
         default_node
       end
     end
-    @right_cell_text += @edit[:adv_search_applied][:text] if x_tree && x_tree[:type] == :cs_filter && @edit && @edit[:adv_search_applied]
+    @right_cell_text += @edit[:adv_search_applied][:text] if x_tree && x_tree[:type] == :configuration_manager_cs_filter && @edit && @edit[:adv_search_applied]
 
     if @edit && @edit.fetch_path(:adv_search_applied, :qs_exp) # If qs is active, save it in history
       x_history_add_item(:id     => x_node,
@@ -558,15 +360,6 @@ class ProviderForemanController < ApplicationController
     end
   end
 
-  def provider_list(id, model)
-    return provider_node(id, model) if id
-    if x_active_tree == :configuration_manager_providers_tree
-      options = {:model => model.to_s}
-      @right_cell_text = _("All %{title} Providers") % {:title => model_to_name(model)}
-      process_show_list(options)
-    end
-  end
-
   def configuration_profile_node(id, model)
     @record = @configuration_profile_record = model ? find_record(ConfigurationProfile, id) : ConfigurationProfile.new
     if @configuration_profile_record.nil?
@@ -592,120 +385,28 @@ class ProviderForemanController < ApplicationController
     end
   end
 
-  def configured_system_list(id, model)
-    return configured_system_node(id, model) if id
-    if x_active_tree == :cs_filter_tree
-      options = {:model => model.to_s}
-      @right_cell_text = _("All %{title} Configured Systems") % {:title => model_to_name(model)}
-      process_show_list(options)
-    end
-  end
-
-  def configured_system_node(id, model)
-    @record = @configured_system_record = find_record(ConfiguredSystem, id)
-    display_node(id, model)
-  end
-
-  def miq_search_node
-    options = {:model => "ConfiguredSystem"}
-    process_show_list(options)
-    @right_cell_text = _("All %{title} Configured Systems") % {:title => ui_lookup(:ui_title => "foreman")}
-  end
-
   def default_node
     return unless x_node == "root"
     if x_active_tree == :configuration_manager_providers_tree
       options = {:model => "ManageIQ::Providers::ConfigurationManager"}
       process_show_list(options)
       @right_cell_text = _("All Configuration Management Providers")
-    elsif x_active_tree == :cs_filter_tree
+    elsif x_active_tree == :configuration_manager_cs_filter_tree
       options = {:model => "ConfiguredSystem"}
       process_show_list(options)
       @right_cell_text = _("All Configured Systems")
     end
   end
 
-  def rendering_objects
-    presenter = ExplorerPresenter.new(
-      :active_tree => x_active_tree,
-      :delete_node => @delete_node,
-    )
-    r = proc { |opts| render_to_string(opts) }
-    return presenter, r
-  end
-
-  def render_form
-    presenter, r = rendering_objects
-    @in_a_form = true
-    presenter.update(:main_div, r[:partial => 'form', :locals => {:controller => 'provider_foreman'}])
-    update_title(presenter)
-    rebuild_toolbars(false, presenter)
-    handle_bottom_cell(presenter, r)
-
-    render :json => presenter.for_render
-  end
-
-  def render_tagging_form
-    return if %w(cancel save).include?(params[:button])
-    @in_a_form = true
-    @right_cell_text = _("Edit Tags")
-    clear_flash_msg
-    presenter, r = rendering_objects
-    update_tagging_partials(presenter, r)
-    update_title(presenter)
-    rebuild_toolbars(false, presenter)
-    handle_bottom_cell(presenter, r)
-
-    render :json => presenter.for_render
-  end
-
-  def update_tree_and_render_list(replace_trees)
-    @explorer = true
-    get_node_info(x_node)
-    presenter, r = rendering_objects
-    replace_explorer_trees(replace_trees, presenter, r)
-
-    presenter.update(:main_div, r[:partial => 'layouts/x_gtl'])
-    rebuild_toolbars(false, presenter)
-    handle_bottom_cell(presenter, r)
-
-    render :json => presenter.for_render
-  end
-
-  def update_title(presenter)
-    @right_cell_text =
-      case action_name
-      when "new"  then _("Add a new Configuration Management Provider")
-      when "edit" then _("Edit Configuration Manager Provider")
-      end
-    presenter[:right_cell_text] = @right_cell_text
-  end
-
-  def replace_right_cell(options = {})
-    replace_trees = options[:replace_trees]
-    return if @in_a_form
-    @explorer = true
-    @in_a_form = false
-    @sb[:action] = nil
-
+  def rebuild_trees(replace_trees)
     trees = {}
     if replace_trees
       trees[:configuration_manager_providers] = build_configuration_manager_tree(:configuration_manager_providers,
                                                                                  :configuration_manager_providers_tree) if replace_trees.include?(:configuration_manager_providers)
-      trees[:cs_filter] = build_configuration_manager_tree(:cs_filter,
-                                                           :cs_filter_tree) if replace_trees.include?(:cs_filter)
+      trees[:configuration_manager_cs_filter] = build_configuration_manager_tree(:configuration_manager_cs_filter,
+                                                           :configuration_manager_cs_filter_tree) if replace_trees.include?(:configuration_manager_cs_filter)
     end
-    record_showing = leaf_record
-    presenter, r = rendering_objects
-    update_partials(record_showing, presenter, r)
-    replace_search_box(presenter, r)
-    handle_bottom_cell(presenter, r)
-    replace_trees_by_presenter(presenter, trees)
-    rebuild_toolbars(record_showing, presenter)
-    presenter[:right_cell_text] = @right_cell_text
-    presenter[:osf_node] = x_node  # Open, select, and focus on this node
-
-    render :json => presenter.for_render
+    trees
   end
 
   def leaf_record
@@ -736,145 +437,34 @@ class ProviderForemanController < ApplicationController
     node
   end
 
-  def apply_node_search_text
-    setup_search_text_for_node
-    previous_nodetype = search_text_type(@sb[:foreman_search_text][:previous_node])
-    current_nodetype  = search_text_type(@sb[:foreman_search_text][:current_node])
-
-    @sb[:foreman_search_text]["#{previous_nodetype}_search_text"] = @search_text
-    @search_text = @sb[:foreman_search_text]["#{current_nodetype}_search_text"]
-    @sb[:foreman_search_text]["#{x_active_accord}_search_text"] = @search_text
-  end
-
-  def setup_search_text_for_node
-    @sb[:foreman_search_text] ||= {}
-    @sb[:foreman_search_text][:current_node] ||= x_node
-    @sb[:foreman_search_text][:previous_node] = @sb[:foreman_search_text][:current_node]
-    @sb[:foreman_search_text][:current_node] = x_node
-  end
-
   def update_partials(record_showing, presenter, r)
     if record_showing && valid_configured_system_record?(@configured_system_record)
       get_tagdata(@record)
       presenter.hide(:form_buttons_div)
-      path_dir = "provider_foreman"
+      path_dir = controller_name
       presenter.update(:main_div, r[:partial => "#{path_dir}/main",
-                                    :locals  => {:controller => 'provider_foreman'}])
+                                    :locals  => {:controller => controller_name}])
     elsif @in_a_form
-      partial_locals = {:controller => 'provider_foreman'}
+      partial_locals = {:controller => controller_name}
       @right_cell_text =
-        if @sb[:action] == "provider_foreman_add_provider"
-          _("Add a new Configuration Manager Provider")
-        elsif @sb[:action] == "provider_foreman_edit_provider"
-          # set the title based on the configuration manager provider type
-          _("Edit Configuration Manager Provider")
+        if @sb[:action] == "#{controller_name}_add_provider"
+          _("Add a new Configuration Management Provider")
+        elsif @sb[:action] == "#{controller_name}_edit_provider"
+          _("Edit Provider")
         end
       partial = 'form'
       presenter.update(:main_div, r[:partial => partial, :locals => partial_locals])
     elsif valid_configuration_profile_record?(@configuration_profile_record)
       presenter.hide(:form_buttons_div)
       presenter.update(:main_div, r[:partial => "configuration_profile",
-                                    :locals  => {:controller => 'provider_foreman'}])
+                                    :locals  => {:controller => controller_name}])
     else
       presenter.update(:main_div, r[:partial => 'layouts/x_gtl'])
     end
   end
 
-  def replace_search_box(presenter, r)
-    # Replace the searchbox
-    presenter.replace(:adv_searchbox_div,
-                      r[:partial => 'layouts/x_adv_searchbox'])
-
-    presenter[:clear_gtl_list_grid] = @gtl_type && @gtl_type != 'list'
-  end
-
-  def handle_bottom_cell(presenter, r)
-    # Handle bottom cell
-    if @pages || @in_a_form
-      if @pages && !@in_a_form
-        @ajax_paging_buttons = true
-        if @sb[:action] && @record # Came in from an action link
-          presenter.update(:paging_div, r[:partial => 'layouts/x_pagingcontrols',
-                                          :locals  => {:action_url    => @sb[:action],
-                                                       :action_method => @sb[:action],
-                                                       :action_id     => @record.id}])
-        else
-          presenter.update(:paging_div, r[:partial => 'layouts/x_pagingcontrols'])
-        end
-        presenter.hide(:form_buttons_div).show(:pc_div_1)
-      elsif @in_a_form
-        presenter.hide(:pc_div_1).show(:form_buttons_div)
-      end
-      presenter.show(:paging_div)
-    else
-      presenter.hide(:paging_div)
-    end
-  end
-
-  def rebuild_toolbars(record_showing, presenter)
-    if configuration_profile_summary_tab_selected?
-      center_tb = "blank_view_tb"
-      record_showing = true
-    end
-
-    if !@in_a_form && !@sb[:action]
-      center_tb ||= center_toolbar_filename
-      c_tb = build_toolbar(center_tb)
-
-      v_tb = if record_showing
-               build_toolbar("x_summary_view_tb")
-             else
-               build_toolbar("x_gtl_view_tb")
-             end
-    end
-
-    h_tb = build_toolbar("x_history_tb") unless @in_a_form
-
-    presenter.reload_toolbars(:history => h_tb, :center => c_tb, :view => v_tb)
-
-    presenter.set_visibility(h_tb.present? || c_tb.present? || v_tb.present?, :toolbar)
-
-    presenter[:record_id] = @record.try(:id)
-
-    # Hide/show searchbox depending on if a list is showing
-    presenter.set_visibility(display_adv_searchbox, :adv_searchbox_div)
-    presenter[:clear_search_toggle] = clear_search_status
-
-    presenter.hide(:blocker_div) unless @edit && @edit[:adv_search_open]
-    presenter.hide(:quicksearchbox)
-    presenter[:hide_modal] = true
-
-    presenter.lock_tree(x_active_tree, @in_a_form)
-  end
-
-  def display_adv_searchbox
-    !(@configured_system_record ||
-      @in_a_form ||
-      configuration_profile_summary_tab_selected?)
-  end
-
-  def configuration_profile_summary_tab_selected?
+  def group_summary_tab_selected?
     @configuration_profile_record && @sb[:active_tab] == 'summary'
-  end
-
-  def construct_edit
-    @edit ||= {}
-    @edit[:current] = {:name       => @provider_cfgmgmt.name,
-                       :provtype   => model_to_name(@provider_cfgmgmt.type),
-                       :url        => @provider_cfgmgmt.url,
-                       :verify_ssl => @provider_cfgmgmt.verify_ssl}
-    @edit[:new] = {:name       => params[:name],
-                   :provtype   => params[:provtype],
-                   :url        => params[:url],
-                   :verify_ssl => params[:verify_ssl]}
-  end
-
-  def breadcrumb_name(_model)
-    "#{ui_lookup(:ui_title => 'foreman')} #{ui_lookup(:model => 'ExtManagementSystem')}"
-  end
-
-  def tagging_explorer_controller?
-    @explorer
   end
 
   def active_tab_configured_systems?
@@ -953,28 +543,10 @@ class ProviderForemanController < ApplicationController
     options[:dbname] = case x_active_accord
                        when :configuration_manager_providers
                          options[:model] && options[:model] == 'ConfiguredSystem' ? :cm_configured_systems : :cm_providers
-                       when :cs_filter
+                       when :configuration_manager_cs_filter
                          :cm_configured_systems
                        end
     super
-  end
-
-  def find_record(model, id)
-    raise _("Invalid input") unless is_integer?(from_cid(id))
-    begin
-      record = model.where(:id => from_cid(id)).first
-    rescue ActiveRecord::RecordNotFound, StandardError => ex
-      if @explorer
-        self.x_node = "root"
-        add_flash(ex.message, :error, true)
-        session[:flash_msgs] = @flash_array.dup
-      end
-    end
-    record
-  end
-
-  def title
-    _("Providers")
   end
 
   menu_section :conf
