+++ app/views/automation_manager/_main.html.haml
@@ -3,6 +3,6 @@
   .col-md-12.col-lg-6
     = render :partial => "shared/summary/textual", :locals => {:title => _("Properties"), :items => textual_group_properties}
   .col-md-12.col-lg-6
-    = render :partial => "shared/summary/textual", :locals => {:title => _("Tenancy"), :items => textual_group_tenancy}
     = render :partial => "shared/summary/textual_tags", :locals => {:title => _("Smart Management"), :items => textual_group_tags}
-
+  .col-md-12.col-lg-6
+    = render :partial => "shared/summary/textual", :locals => {:title => _("Relationships"), :items => textual_group_relationships}
